Fix all lint errors

+++ commands/set_persistent_message.js
@@ -34,9 +34,10 @@ async function checkPersistentDeletion(guildStore, channelId, messageId) {
     persistChannel.active = false;
     await guild.save();
 
-    channelId = client.channels.cache.get(channelId);
-    if (channelId.permissionsFor(client.user.id).has(PermissionFlagsBits.SendMessages)) {
-        channelId.send(`I have detected that a moderator deleted the persistent message set for this channel, and as such I have deactivated it. To reactivate it, a moderator can run ${cmds.set_persistent_message.mention}.`);
+    const channel = client.channels.cache.get(channelId);
+    if (channel.isSendable()) {
+        // @ts-ignore
+        channel.send(`I have detected that a moderator deleted the persistent message set for this channel, and as such I have deactivated it. To reactivate it, a moderator can run ${cmds?.set_persistent_message?.mention ?? "the persistent message command"}.`).catch(() => {});
     }
 }
 
@@ -92,8 +93,15 @@ module.exports = {
         guild.persistence.get(cmd.channel.id).active=cmd.options.getBoolean("active");
 
         if(cmd.options.getString("content")!==null) guild.persistence.get(cmd.channel.id).content=cmd.options.getString("content");
-        if(cmd.channel.permissionsFor(client.user.id).has(PermissionFlagsBits.ManageWebhooks)&&cmd.channel.permissionsFor(client.user.id).has(PermissionFlagsBits.ReadMessageHistory)&&cmd.channel.permissionsFor(client.user.id).has(PermissionFlagsBits.ManageMessages)){
-            cmd.followUp(`I have set your settings for this channel's persistent messages.`);
+        const botPerms = cmd.channel.permissionsFor(client.user.id);
+        if (
+            botPerms?.has(PermissionFlagsBits.ManageWebhooks) && 
+            botPerms.has(PermissionFlagsBits.ReadMessageHistory) && 
+            botPerms.has(PermissionFlagsBits.ManageMessages) &&
+            "fetchWebhooks" in cmd.channel &&
+            "createWebhook" in cmd.channel
+        ) {
+            await cmd.followUp({ content: `I have set your settings for this channel's persistent messages.` });
 
             var resp = {
                 "content": await censor(guild.persistence.get(cmd.channel.id).content),
@@ -103,41 +111,39 @@ module.exports = {
 
             // If they just set it to active, send the message
             if (guild.persistence.get(cmd.channel.id).active) {
-                // Discord server name edge case
+                // Discord server name edge case - TODO: functionize to use elsewhere
                 if (resp?.username?.toLowerCase().includes("discord")) {
-                    resp.username = "[SERVER]"
+                    resp.username = "[SERVER]";
                 }
-                var hook=await cmd.channel.fetchWebhooks();
-                hook=hook.find(h=>h.token);
-                if(hook){
-                    hook.send(resp).then(d=>{
-                        guild.persistence.get(cmd.channel.id).lastPost=d.id;
-                    });
+                let webhooks = await cmd.channel.fetchWebhooks();
+                let hook = webhooks.find(h => h.token);
+                if (!hook) {
+                    const channel = await client.channels.fetch(cmd.channel.id);
+                    hook = (channel && "createWebhook" in channel) 
+                        ? await channel.createWebhook({
+                            name: config.name,
+                            avatar: config.pfp
+                        }) 
+                        : null;
                 }
-                else{
-                    client.channels.cache.get(cmd.channel.id).createWebhook({
-                        name: config.name,
-                        avatar: config.pfp
-                    }).then(d=>{
-                        d.send(resp).then(d=>{
-                            guild.persistence.get(cmd.channel.id).lastPost=d.id;
-                        });
-                    });
+                if (hook) {
+                    const sent = await hook.send(resp);
+                    guild.persistence.get(cmd.channel.id).lastPost = sent.id;
                 }
             }
         }
-        else{
+        else {
             cmd.followUp(`I need to be able to read message history, delete messages, and manage webhooks to use persistent messages. Without these permissions I cannot manage persistent messages here.`);
             guild.persistence.get(cmd.channel.id).active=false;
         }
 
-        guild.save();
+        await guild.save();
 	},
 
     /** 
      * @param {import('discord.js').Message} msg 
-     * @param {GuildDoc} guildStore 
-     * @param {GuildUserDoc} guildUserStore 
+     * @param {import("./modules/database.js").GuildDoc} guildStore 
+     * @param {import("./modules/database.js").GuildUserDoc} guildUserStore 
      * */
     async [Events.MessageCreate] (msg, context, guildStore, guildUserStore) {
         if (msg.webhookId) return;
@@ -148,8 +154,14 @@ module.exports = {
         const guild = guildStore;
 
         // Persistent messages, if the server has them enabled
-        if(guild.persistence[msg.channel.id]?.active) {
-            if(msg.channel.permissionsFor(client.user.id).has(PermissionFlagsBits.ManageWebhooks) && msg.channel.permissionsFor(client.user.id).has(PermissionFlagsBits.ManageMessages) && msg.channel.permissionsFor(client.user.id).has(PermissionFlagsBits.ReadMessageHistory)){
+        if(
+            guild.persistence[msg.channel.id]?.active && 
+            "permissionsFor" in msg.channel &&
+            "fetchWebhooks" in msg.channel &&
+            "createWebhook" in msg.channel
+        ) {
+            const permissions = msg.channel.permissionsFor(client.user.id);
+            if(permissions?.has(PermissionFlagsBits.ManageWebhooks) && permissions.has(PermissionFlagsBits.ManageMessages) && permissions.has(PermissionFlagsBits.ReadMessageHistory)){
                 if(guild.persistence[msg.channel.id].lastPost){
                     msg.channel.messages.fetch(guild.persistence[msg.channel.id].lastPost).then(mes => {
                         mes?.delete()?.catch(e=>{});
@@ -160,16 +172,20 @@ module.exports = {
                     "avatarURL":msg.guild.iconURL(),
                     "username":msg.guild.name
                 };
-                let hook = await msg.channel.fetchWebhooks();
-                hook = hook.find(h => h.token);
+                let webhooks = await msg.channel.fetchWebhooks();
+                let hook = webhooks.find(h => h.token);
 
                 if (!hook) {
                     const channel = await client.channels.fetch(msg.channel.id);
-                    hook = await channel.createWebhook({
-                        name: config.name,
-                        avatar: config.pfp
-                    });
+                    hook = (channel && "createWebhook" in channel) 
+                        ? await channel.createWebhook({
+                            name: config.name,
+                            avatar: config.pfp
+                        }) 
+                        : null;
                 }
+                if (!hook) return;
+
                 const newMessage = await hook.send(resp)
                 
                 await guildByObj(msg.guild, {
@@ -178,7 +194,8 @@ module.exports = {
 
             }
             else {
-                if(msg.channel.permissionsFor(client.user.id).has(PermissionFlagsBits.SendMessages)){
+                if(msg.channel.isSendable()){
+                    // @ts-ignore
                     msg.channel.send(`I do not have sufficient permissions to manage persistent messages for this channel. Please make sure I can manage webhooks, read message history, and delete messages and then run ${cmds.set_persistent_message.mention}.`);
                 }
